requirements.txt
google-cloud-datacatalog
google-cloud-bigquery
google-auth
